Extract code to verify ISO session transcript
