+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/SessionTranscriptCalculator.kt
@@ -1,0 +1,105 @@
+package at.asitplus.wallet.lib.openid
+
+import at.asitplus.dcapi.DCAPIHandover
+import at.asitplus.dcapi.OpenID4VPDCAPIHandoverInfo
+import at.asitplus.iso.OpenId4VpHandover
+import at.asitplus.iso.OpenId4VpHandoverInfo
+import at.asitplus.iso.SessionTranscript
+import at.asitplus.iso.serializeOrigin
+import at.asitplus.iso.sha256
+import at.asitplus.openid.ResponseParametersFrom
+import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
+import at.asitplus.wallet.lib.agent.KeyMaterial
+import at.asitplus.wallet.lib.extensions.sessionTranscriptThumbprint
+import kotlinx.serialization.encodeToByteArray
+
+internal fun interface SessionTranscriptCalculator {
+    operator fun invoke(
+        input: ResponseParametersFrom,
+        clientId: String?,
+        expectedNonce: String,
+        hasBeenEncrypted: Boolean,
+        responseUrl: String?,
+        clientIdRequired: Boolean,
+        origin: String?,
+    ): SessionTranscript
+}
+
+/** Calculates the ISO session transcript for URL transport, i.e. from [OpenId4VpVerifier]. */
+internal class UrlSessionTranscriptCalculator(
+    private val decryptionKeyMaterial: KeyMaterial
+) : SessionTranscriptCalculator {
+    override fun invoke(
+        input: ResponseParametersFrom,
+        clientId: String?,
+        expectedNonce: String,
+        hasBeenEncrypted: Boolean,
+        responseUrl: String?,
+        clientIdRequired: Boolean,
+        origin: String?
+    ): SessionTranscript {
+        require((!clientIdRequired || clientId != null)) { "Missing required parameter: clientId" }
+        require(responseUrl != null) { "Missing required parameter: responseUrl" }
+        require(input.originalResponseParameters !is ResponseParametersFrom.DcApi) {
+            "DCAPI verification is not supported, use DcApiVerifier"
+        }
+
+        return SessionTranscript.forOpenId(
+            OpenId4VpHandover(
+                type = OpenId4VpHandover.TYPE_OPENID4VP,
+                hash = coseCompliantSerializer.encodeToByteArray<OpenId4VpHandoverInfo>(
+                    OpenId4VpHandoverInfo(
+                        clientId = clientId,
+                        nonce = expectedNonce,
+                        jwkThumbprint = if (hasBeenEncrypted) {
+                            decryptionKeyMaterial.jsonWebKey.sessionTranscriptThumbprint()
+                        } else null,
+                        responseUrl = responseUrl,
+                    )
+                ).sha256(),
+            )
+        )
+    }
+}
+
+/** Calculates the ISO session transcript for DCAPI transport, i.e. from [DcApiVerifier]. */
+internal class DcApiSessionTranscriptCalculator(
+    private val decryptionKeyMaterial: KeyMaterial
+) : SessionTranscriptCalculator {
+    override fun invoke(
+        input: ResponseParametersFrom,
+        clientId: String?,
+        expectedNonce: String,
+        hasBeenEncrypted: Boolean,
+        responseUrl: String?,
+        clientIdRequired: Boolean,
+        origin: String?
+    ): SessionTranscript {
+        require((!clientIdRequired || clientId != null)) { "Missing required parameter: clientId" }
+        require(responseUrl != null) { "Missing required parameter: responseUrl" }
+        require(input.originalResponseParameters is ResponseParametersFrom.DcApi) {
+            "Unsupported response mechanism: ${input.originalResponseParameters}"
+        }
+        require(origin != null) { "Missing required parameter: origin" }
+        val serializedOrigin = requireNotNull(origin.serializeOrigin()) {
+            "Invalid parameter: origin"
+        }
+        return SessionTranscript.forDcApi(
+            DCAPIHandover(
+                type = DCAPIHandover.TYPE_OPENID4VP,
+                hash = coseCompliantSerializer.encodeToByteArray<OpenID4VPDCAPIHandoverInfo>(
+                    OpenID4VPDCAPIHandoverInfo(
+                        // Device signatures are bound to the HTML-serialized origin used by OpenID4VP/DCAPI.
+                        // Hashing the raw URL would make `https://example.com/` differ from `https://example.com`.
+                        origin = serializedOrigin,
+                        nonce = expectedNonce,
+                        jwkThumbprint = if (hasBeenEncrypted) {
+                            decryptionKeyMaterial.jsonWebKey.sessionTranscriptThumbprint()
+                        } else null,
+                    )
+                ).sha256(),
+            )
+        )
+    }
+
+}
